backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
supabase==1.0.4
pinecone==7.3.0
openai==1.3.0
python-dotenv==1.0.0
numpy==1.24.3
sentence-transformers==2.2.2
